Utilized the parseInt method within the Integer class to return the final value

+++ src/main/java/com/example/SquareDigit.java
@@ -26,7 +26,7 @@ public class SquareDigit {
 
             double digit = Math.pow(Character.getNumericValue(numberStr.charAt(i)), 2);
 
-            result.append(digit);
+            result.append((int) digit);
 
 
         }
